+++ IMPLC.Core/Repository/DeviceRepoFactory.cs
@@ -2,7 +2,7 @@
 {
     public class DeviceRepoFactory
     {
-        public static IDeviceRepo MakeDeviceRepo(Device device, short length)
+        public static IDeviceRepo MakeDeviceRepo(Device device, int length)
         {
             if (device.IsBitDevice())
                 return new BitDeviceRepo(length);
